libmicrohttpd API fix obvious on 32bit only
+ It's uint64_t for the stream position for dynamic replies, this makes a
. difference on 32bit systems... Thanks to Daniel for noticing.

Tim


git-svn-id: https://lechuck.informatik.rwth-aachen.de:2342/svn/fawkes/trunk@1977 fc82b16a-671c-0410-8732-9903a8d2dd0d

+++ src/plugins/webview/request_dispatcher.cpp
@@ -79,7 +79,7 @@ WebRequestDispatcher::process_request_cb(void *callback_data,
  */
 int
 WebRequestDispatcher::dynamic_reply_data_cb(void *reply,
-					    size_t pos, char *buf, int max)
+					    uint64_t pos, char *buf, int max)
 {
   DynamicWebReply *dreply = static_cast<DynamicWebReply *>(reply);
   return dreply->next_chunk(pos, buf, max);
